Overhaul documentation.
* More documentation pages are added. Everything is no-longer in the
main page.
* The documentation now includes the library version it's for.
* A few irrelevant Qt classes were removed from the documentation.
* A link to an example project was added.
* Turned an enum in QDiscordGame into a scoped enum.
* Made sure a majority of the code fits on an 80-character-wide terminal.

+++ QDiscord/src/qdiscord.d/qdiscordwscomponent.cpp
@@ -105,7 +105,7 @@ void QDiscordWsComponent::setStatus(bool idle, QDiscordGame game)
 		QJsonObject gameObject;
 		gameObject["name"] = game.name();
 		gameObject["url"] = game.url()==""?QJsonValue():game.url();
-		gameObject["type"] = game.type();
+		gameObject["type"] = static_cast<int>(game.type());
 		presenceObject["game"] = gameObject;
 	}
 	else
